refactor(DhanGuruPane.jsx): dhan guru shortcut code placement
