+++ www/main/navigator/misc/components/InsertPane.jsx
@@ -1,29 +1,21 @@
-import React, { useRef, useEffect, useState } from 'react';
+import React, { useRef, useEffect } from 'react';
 import PropTypes from 'prop-types';
 import { ipcRenderer } from 'electron';
 import { useStoreState, useStoreActions } from 'easy-peasy';
-import insertSlide from '../../../common/constants/slidedb';
 
 const remote = require('@electron/remote');
-const tingle = require('../../../../assets/js/vendor/tingle');
 
-const { gurus } = insertSlide.dropdownStrings;
 const { i18n } = remote.require('./app');
 const analytics = remote.getGlobal('analytics');
 
 export const InsertPane = ({ className }) => {
-  const { isMiscSlide, isMiscSlideGurmukhi, miscSlideText, isAnnoucement, shortcuts } =
-    useStoreState((state) => state.navigator);
-  const {
-    setIsMiscSlide,
-    setIsMiscSlideGurmukhi,
-    setMiscSlideText,
-    setIsAnnoucement,
-    setShortcuts,
-  } = useStoreActions((state) => state.navigator);
+  const { isMiscSlide, isMiscSlideGurmukhi, miscSlideText, isAnnoucement } = useStoreState(
+    (state) => state.navigator,
+  );
+  const { setIsMiscSlide, setIsMiscSlideGurmukhi, setMiscSlideText, setIsAnnoucement } =
+    useStoreActions((state) => state.navigator);
 
   const inputRef = useRef(null);
-  const [isModalOpen, setIsModalOpen] = useState(false);
 
   const addMiscSlide = (givenText) => {
     if (!isMiscSlide) {
@@ -48,74 +40,6 @@ export const InsertPane = ({ className }) => {
     }
   };
 
-  const addDhanGuruSlide = (e) => {
-    if (isAnnoucement) {
-      setIsAnnoucement(false);
-    }
-    if (typeof e === 'object') {
-      addMiscSlide(e.target.value);
-      analytics.trackEvent('display', 'dhanguru-slide', e.target.value);
-    } else {
-      addMiscSlide(e);
-      analytics.trackEvent('display', 'dhanguru-slide', e);
-    }
-  };
-
-  let slidePage = `<h1 class = "modalTitle">${i18n.t('INSERT.INSERT_DHAN_SLIDE')}</h1>
-  <div class="btn-group" id = "btn-group">`;
-  gurus.forEach((guru, index) => {
-    slidePage += `<button class="guru" id="guru${index}">${i18n.t(
-      `INSERT.DHAN_GURU.${guru}`,
-    )}</button>`;
-  });
-  slidePage += `</div>`;
-
-  const showDhanGuruModal = () => {
-    if (!isModalOpen) {
-      const modal = new tingle.Modal({
-        footer: true,
-        stickyFooter: false,
-        closeMethods: ['overlay', 'button', 'escape'],
-        onClose() {
-          modal.modal.classList.remove('tingle-modal--visible');
-          setIsModalOpen(false);
-          modal.destroy();
-        },
-        beforeClose() {
-          return true; // close the modal
-        },
-      });
-      if (!modal.isOpen()) {
-        setIsModalOpen(true);
-        const buttonOnClick = () => {
-          gurus.forEach((guru, index) => {
-            document.querySelector(`#guru${index}`).onclick = () => {
-              addDhanGuruSlide(insertSlide.slideStrings.dhanguruStrings[index]);
-              modal.close();
-              setIsModalOpen(false);
-              modal.destroy();
-            };
-          });
-        };
-
-        // sets the default page to Dhan Guru slide page
-        modal.setContent(slidePage);
-        buttonOnClick();
-        modal.open();
-      }
-    }
-  };
-
-  useEffect(() => {
-    if (shortcuts.openDhanGuruSlide) {
-      showDhanGuruModal();
-      setShortcuts({
-        ...shortcuts,
-        openDhanGuruSlide: false,
-      });
-    }
-  }, [shortcuts]);
-
   useEffect(() => {
     if (isMiscSlide) {
       ipcRenderer.send('show-misc-text', {
@@ -128,22 +52,6 @@ export const InsertPane = ({ className }) => {
 
   return (
     <ul className={`list-of-items ${className}`}>
-      <li>
-        <a>
-          <i className="fa fa-circle-o list-icon" />
-          <label>{i18n.t('INSERT.ADD_DHAN_GURU')} </label>
-          <select onChange={addDhanGuruSlide}>
-            <option value="" disabled>
-              {i18n.t('INSERT.SELECT')}
-            </option>
-            {gurus.map((value, index) => (
-              <option value={insertSlide.slideStrings.dhanguruStrings[index]} key={index}>
-                {i18n.t(`INSERT.DHAN_GURU.${value}`)}
-              </option>
-            ))}
-          </select>
-        </a>
-      </li>
       <li className="announcement-box">
         <header>
           <i className="fa fa-bullhorn list-icon" />
